[VPR] Compiler warning fix

+++ vpr/src/device/rr_spatial_lookup.cpp
@@ -94,8 +94,8 @@ void RRSpatialLookup::add_node(RRNodeId node,
     if ((x >= int(rr_node_indices_[type].dim_size(0)))
         || (y >= int(rr_node_indices_[type].dim_size(1)))
         || (size_t(side) >= rr_node_indices_[type].dim_size(2))) {
-        rr_node_indices_[type].resize({std::max(int(rr_node_indices_[type].dim_size(0)), x + 1),
-                                       std::max(int(rr_node_indices_[type].dim_size(1)), y + 1),
+        rr_node_indices_[type].resize({std::max(rr_node_indices_[type].dim_size(0), size_t(x) + 1),
+                                       std::max(rr_node_indices_[type].dim_size(1), size_t(y) + 1),
                                        std::max(rr_node_indices_[type].dim_size(2), size_t(side) + 1)});
     }
